back-end/requirements.txt
arrow==0.15.6
autopep8==1.5.2
biopython==1.76
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.3
pandas==1.0.3
path==13.2.0
pycodestyle==2.5.0
python-dateutil==2.8.1
pytz==2020.1
rope==0.16.0
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==1.0.1
zipp==3.1.0
dna-features-viewer==0.1.3
